requirements.txt
Flask>=3.0.0
gunicorn>=21.0.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
google-cloud-storage>=2.14.0
python-dotenv>=1.0.0
schedule>=1.2.0
